+++ +sw_tests/+unit_tests/unittest_super.m
@@ -13,13 +13,7 @@ classdef unittest_super < matlab.mock.TestCase
             path = fullfile(testCase.get_unit_test_dir(), 'Figure', filename);
             obj = openfig(path, 'invisible');
         end
-        function verify_obj(testCase, expected_obj, actual_obj, rel_tol, abs_tol)
-            if nargin < 5
-                abs_tol = 0;
-            end
-            if nargin < 4
-                rel_tol = 1e-10;
-            end
+        function verify_obj(testCase, expected_obj, actual_obj, varargin)
             testCase.assertClass(actual_obj, class(expected_obj));
             all_fieldnames = fieldnames(expected_obj);
             if isa(expected_obj, 'struct')
@@ -33,24 +27,46 @@ classdef unittest_super < matlab.mock.TestCase
                 expected_value = expected_obj.(field{:});
                 actual_value = actual_obj.(field{:});
                 if isstruct(expected_value)
-                    testCase.verify_obj(expected_value, actual_value, rel_tol, abs_tol);
+                    testCase.verify_obj(expected_value, actual_value, varargin{:});
                 else
-                    testCase.verify_val(expected_value, actual_value, rel_tol, abs_tol, field{:});
+                    testCase.verify_val(expected_value, actual_value, ...
+                        'field', field{:}, varargin{:});
                 end
             end
         end
-        function verify_val(testCase, expected_val, actual_val, rel_tol, abs_tol, field)
-            if nargin < 5
-                abs_tol = 0;
-            end
-            if nargin < 4
-                rel_tol = 1e-10;
-            end
+        function verify_val(testCase, expected_val, actual_val, varargin)
             import matlab.unittest.constraints.IsEqualTo
             import matlab.unittest.constraints.RelativeTolerance
             import matlab.unittest.constraints.AbsoluteTolerance
-            theseBounds = RelativeTolerance(rel_tol) | AbsoluteTolerance(abs_tol);
-            testCase.verifyThat(actual_val, IsEqualTo(expected_val, 'Within', theseBounds), field);
+            
+            field = "";
+            abs_tol = 10*eps; % abs_tol comparsion is default
+            rel_tol = 0;
+            for iarg = 1:2:numel(varargin)
+                switch varargin{iarg}
+                    case 'abs_tol'
+                        abs_tol = varargin{iarg + 1};
+                    case 'rel_tol'
+                        if expected_val ~= 0
+                            rel_tol = varargin{iarg + 1};
+                        end
+                    case 'field' 
+                        field = varargin{iarg + 1};
+                end
+            end
+            bounds = RelativeTolerance(rel_tol) | AbsoluteTolerance(abs_tol);
+            testCase.verifyThat(actual_val, ...
+                IsEqualTo(expected_val, 'Within', bounds), field);
+        end
+        function verify_spinw_matrix(testCase, expected_matrix, actual_matrix, varargin)
+            % compare excl. color (which is randomly generated)
+            testCase.verify_val(rmfield(expected_matrix, 'color'), ...
+                rmfield(actual_matrix, 'color'), varargin{:})
+            % check size and data type of color
+            testCase.assertEqual(size(actual_matrix.color), ...
+             [3, size(expected_matrix.mat, 3)]);
+            testCase.assertTrue(isa(actual_matrix.color, ...
+                class(expected_matrix.color)));
         end
     end
 end