+++ swfiles/@spinw/addmatrix.m
@@ -81,8 +81,7 @@ inpForm.soft   = {true       true       true    true    };
 newMat = sw_readparam(inpForm, varargin{:});
 
 if ~isnumeric(newMat.value) && ~isa(newMat.value,'sym')
-    warning('spinw:addmatrix:WrongInput','Matrix value has to be numeric or symbolic variable!')
-    return
+    error('spinw:addmatrix:WrongInput','Matrix value has to be numeric or symbolic variable!')
 end
 
 if ~isempty(newMat.value)