+++ CHANGELOG.rst
@@ -18,3 +18,11 @@ Bug Fixes
 - A ``spinw`` object can now be correctly created from a structure figure
 - ``.cif`` files with a mixture of tabs and spaces or containing a ``?``
   in the comments can now be read correctly
+- Rotation matrix ``rotC``  in ``addtwin`` is now required to be a valid
+  rotation or reflection matrix.
+- Spin of atom in ``addatom`` must have ``S>=0``.
+- Anisotropic g-tensor in ``addg`` must be physically valid - i.e.
+  :math:`g^\dagger.g` must be a symmetric positive definite matrix.
+- Fix bug in addcoupling that did not allow user to supply 'atom' with
+  numeric array of atom indices (previously only worked for string or
+  cell of strings corresponding to atom labels).